requirements.txt
numpy>=1.24.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
requests>=2.31.0
pypdf>=3.0.0